+++ .eslintrc.json
@@ -1,23 +1,23 @@
-// {
-//   "env": {
-//     "browser": true,
-//     "es2021": true,
-//     "node": true
-//   },
-//   "extends": "airbnb-base",
-//   "overrides": [],
-//   "parserOptions": {
-//     "ecmaVersion": "latest",
-//     "sourceType": "module"
-//   },
-//   "rules": {
-//     "linebreak-style": ["warn"],
-//     "no-console": "off",
-//     "no-param-reassign": ["warn", { "props": false }],
-//     "import/extensions": ["warn", "always"],
-//     "class-methods-use-this": "off",
-//     "max-params": ["warn", 2],
-//     "max-depth": ["warn", 1],
-//     "max-lines-per-function": ["warn", 12]
-//   }
-// }
+{
+  "env": {
+    "browser": true,
+    "es2021": true,
+    "node": true
+  },
+  "extends": "airbnb-base",
+  "overrides": [],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "rules": {
+    "linebreak-style": ["warn"],
+    "no-console": "off",
+    "no-param-reassign": ["warn", { "props": false }],
+    "import/extensions": ["warn", "always"],
+    "class-methods-use-this": "off",
+    "max-params": ["warn", 2],
+    "max-depth": ["warn", 1],
+    "max-lines-per-function": ["warn", 12]
+  }
+}
