+++ src/index.html
@@ -8,7 +8,7 @@
     <style type="text/css">
       @import url('css/index.css');
     </style>
-    <script type="module" src="./WebApp.js" defer></script>
+    <script type="module" src="webView/WebApp.js" defer></script>
   </head>
   <body>
     <header>
@@ -23,92 +23,9 @@
           <button id="purchasing-button" type="submit">구입</button>
         </form>
       </section>
-
-      <!-- 🏖️ section2 -->
-
-      <!-- <section id="lottos-issued">
-        <p>총 7개를 구매하였습니다.</p>
-        <div id="ticket-list">
-          <div class="ticket"><span>🎟️</span> 12,19,22,23,28,37</div>
-          <div class="ticket"><span>🎟️</span> 12,19,22,23,28,37</div>
-          <div class="ticket"><span>🎟️</span> 12,19,22,23,28,37</div>
-          <div class="ticket"><span>🎟️</span> 12,19,22,23,28,37</div>
-        </div>
-      </section> -->
-
-      <!-- 🏖️ section3 -->
-      <!-- <section id="winning-lotto">
-        <p class="message">지난 주 당첨번호 6개와 보너스 번호 1개를 입력해주세요.</p>
-        <form action="">
-          <div id="number-message" class="message">
-            <span>당첨 번호</span>
-            <span>보너스 번호</span>
-          </div>
-          <div id="number-box">
-            <div id="winning-numbers">
-              <input type="text" minlength="1" maxlength="2" />
-              <input type="text" minlength="1" maxlength="2" />
-              <input type="text" minlength="1" maxlength="2" />
-              <input type="text" minlength="1" maxlength="2" />
-              <input type="text" minlength="1" maxlength="2" />
-              <input type="text" minlength="1" maxlength="2" />
-            </div>
-            <div id="bonus-number">
-              <input type="text" />
-            </div>
-          </div>
-          <button id="result-button" type="submit">결과 확인하기</button>
-        </form>
-      </section> -->
     </div>
     <footer>
       <p id="copyright">Copyright 2023. woowacourse</p>
     </footer>
-
-    <!-- 🏖️ modal -->
-    <!-- <div class="modal-background">
-      <div id="result-modal">
-        <section id="result-title">
-          🏆 당첨 통계 🏆
-          <span class="close-button">X</span>
-        </section>
-        <section id="result-table">
-          <div id="result-head" class="result-row">
-            <div class="result-item">일치 갯수</div>
-            <div class="result-item">당첨금</div>
-            <div class="result-item">당첨 갯수</div>
-          </div>
-          <div class="result-row">
-            <div class="result-item">3개</div>
-            <div class="result-item">5,000</div>
-            <div class="result-item">1개</div>
-          </div>
-          <div class="result-row">
-            <div class="result-item">4개</div>
-            <div class="result-item">50,000</div>
-            <div class="result-item"></div>
-          </div>
-          <div class="result-row">
-            <div class="result-item">5개+보너스볼</div>
-            <div class="result-item"></div>
-            <div class="result-item"></div>
-          </div>
-          <div class="result-row">
-            <div class="result-item"></div>
-            <div class="result-item"></div>
-            <div class="result-item"></div>
-          </div>
-          <div class="result-row">
-            <div class="result-item">6개</div>
-            <div class="result-item">2,000,000,000</div>
-            <div class="result-item"></div>
-          </div>
-          <div id="result-earning">당신의 총 수익률은 %입니다.</div>
-          <form action="" id="retry-form">
-            <button id="retry-button" type="submit">다시 시작하기</button>
-          </form>
-        </section>
-      </div>
-    </div> -->
   </body>
 </html>
