+++ src/webView/WebApp.js
@@ -1,103 +1,36 @@
-import LottoGame from './domain/LottoGame.js';
+import LottoGame from '../domain/LottoGame.js';
+import { createDomWith, appendDomByList, $, all$ } from './domUtils.js';
+import { renderLottoList } from './renderLottoList.js';
+
 const lottoGame = new LottoGame();
 
-// utils
-const $ = (query) => document.querySelector(query);
-
-const all$ = (query) => document.querySelectorAll(query);
-
-const setDomAttributes = ($element) => (options) => {
-  Object.assign($element, options);
-  return $element;
-};
-
-const createDomWith = (tag) => (options) => setDomAttributes(document.createElement(tag))(options);
-
-const appendDomByList = ($parent) => (children) => (handler) => {
-  $parent.append(...children.map(handler));
-};
-
-const appendBySequence = (...domList) => {
-  if (!domList || domList.length <= 1) return;
-  domList.forEach((element, index) => {
-    if (index === domList.length - 1) return;
-    domList[index + 1].appendChild(element);
-  });
-};
-
-// main
-
-const makeTicketDom = (lotto) => {
-  const $ticket = createDomWith('div')({ className: 'ticket' });
-  const $ticketEmoji = document.createElement('span');
-  $ticketEmoji.innerText = '🎟️';
-
-  const $ticketNumbers = document.createElement('span');
-  $ticketNumbers.innerText = `${[...lotto].join(', ')}`;
-
-  $ticket.append($ticketEmoji, $ticketNumbers);
-
-  return $ticket;
-};
-
-const getTicketCount = (lottoList) => {
-  const ticketMessage = document.createElement('p');
-  ticketMessage.textContent = `총 ${lottoList.length}개를 구매하였습니다.`;
-
-  return ticketMessage;
-};
-
-const renderLottoList = (lottoList) => {
-  const $ticketSection = createDomWith('section')({ id: 'lotto-issued' });
-  const $ticketMessage = getTicketCount(lottoList);
-
-  const $ticketContainer = createDomWith('div')({ id: 'ticket-list' });
-
-  appendDomByList($ticketContainer)(lottoList)(makeTicketDom);
-  $ticketSection.append($ticketMessage, $ticketContainer);
-
-  appendBySequence($ticketSection, $('.app-box'));
-};
-
 const renderWinningForm = () => {
-  // section
   const $winningSection = createDomWith('section')({ id: 'winning-lotto' });
 
-  // p
   const $message = createDomWith('p')({
     class: 'message',
     innerText: '지난 주 당첨번호 6개와 보너스 번호 1개를 입력해주세요.',
   });
 
-  // form
   const $form = document.createElement('form');
-  // div
   const $spanContainer = createDomWith('div')({ id: 'number-message', class: 'message' });
 
-  // span
   const $winningSpan = createDomWith('span')({ innerText: '당첨 번호' });
   const $bonusSpan = createDomWith('span')({ innerText: '보너스 번호' });
   $spanContainer.append($winningSpan, $bonusSpan);
 
-  // div
   const $lottoContainer = createDomWith('div')({ id: 'number-box' });
-
-  // div
   const $lottoWinningContainer = createDomWith('div')({ id: 'winning-numbers' });
 
-  // inputS
-  appendDomByList($lottoWinningContainer)(Array.from({ length: 6 }))((elem) => {
-    return createDomWith('input')({ type: 'text', minlength: '1', maxlength: '2' });
-  });
+  appendDomByList($lottoWinningContainer)(Array.from({ length: 6 }))((elem) =>
+    createDomWith('input')({ type: 'text', minlength: '1', maxlength: '2' })
+  );
 
-  // div
   const $lottoBonusContainer = createDomWith('div')({ id: 'bonus-number' });
 
-  // input
   const $bonusInput = createDomWith('input')({ type: 'text' });
   $lottoBonusContainer.appendChild($bonusInput);
 
-  // submitButton
   const $submitButton = createDomWith('button')({
     id: 'result-button',
     type: 'submit',
@@ -111,7 +44,6 @@ const renderWinningForm = () => {
 
   $('#purchasing-button').disabled = 'disabled';
 
-  // addEventListener
   $submitButton.addEventListener('click', (event) => {
     event.preventDefault();
     try {
@@ -119,17 +51,17 @@ const renderWinningForm = () => {
         Number(elem.value)
       );
       const bonusNumber = Number($('#bonus-number>input').value);
-      console.log(winningNumbers, bonusNumber);
       const resultBoard = lottoGame.setWinningLotto(winningNumbers, bonusNumber).getGameResult();
       renderResult(resultBoard);
     } catch (error) {
-      console.dir(error);
+      alert(error);
     }
   });
 };
 
 const renderResult = ({ first, second, third, fourth, fifth, lottoYield }) => {
   const $modalBackground = createDomWith('div')({ className: 'modal-background' });
+
   $modalBackground.innerHTML = `
   <div id="result-modal">
     <section id="result-title">
@@ -177,13 +109,22 @@ const renderResult = ({ first, second, third, fourth, fifth, lottoYield }) => {
   </div>`;
 
   $('.app-box').appendChild($modalBackground);
-  $('.close-button').onclick = (event) => {
-    event.target.parentNode.parentNode.parentNode.remove();
+  $('.close-button').onclick = () => {
+    $('.modal-background').remove();
   };
+
+  window.addEventListener('keydown', function keyHandler(event) {
+    if (event.key === 'Escape') {
+      $('.modal-background').remove();
+      window.removeEventListener('keydown', keyHandler);
+    }
+
+    if (event.key === 'Enter') {
+      window.location.reload();
+    }
+  });
 };
 
-// event : purchasingButton
-const purchasingButton = $('#purchasing-button');
 const purchasingHandler = (event) => {
   event.preventDefault();
   try {
@@ -196,6 +137,9 @@ const purchasingHandler = (event) => {
   }
 };
 
-// new Promise(reject) {if(lottoList...) reject(event.target)}
+const WebApp = () => {
+  const purchasingButton = $('#purchasing-button');
+  purchasingButton.addEventListener('click', purchasingHandler);
+};
 
-purchasingButton.addEventListener('click', purchasingHandler);
+WebApp();
